+++ opencascade-hs/cpp/hs_types.h
@@ -63,6 +63,7 @@ typedef void BRepOffsetAPI_MakeOffsetShape;
 typedef int BRepOffset_Mode;
 typedef int GeomAbs_JoinType;
 typedef void GProp_GProps;
+typedef void Bnd_Box;
 #define Handle(X) void
 #define ARRAY_1(X) void
 #else // __cplusplus