+++ opencascade-hs/src/OpenCascade/Bnd/Types.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.Bnd.Types
+( Box
+) where
+
+data Box